+++ asimtools/asimmodules/singlepoint.py
@@ -32,7 +32,7 @@ def singlepoint(
     """
     calc = load_calc(calc_id)
     atoms = get_atoms(**image)
-    atoms.set_calculator(calc)
+    atoms.calc = calc
 
     if prefix is not None:
         prefix = prefix + '_'